@@ -7,68 +7,79 @@ readme = "README.md"
 packages = [{include = "fastapi_template"}]
 
 [tool.poetry.dependencies]
-python = "^3.11"
-fastapi = "^0.103.1"
-uvicorn = {extras = ["standard"], version = "^0.23.2"}
-gunicorn = "^21.2.0"
-orjson = "^3.9.7"
-pydantic-settings = "^2.0.3"
-environs = "^9.5.0"
+python = "^3.12"
+fastapi = "^0.111.0"
+uvicorn = {extras = ["standard"], version = "^0.30.1"}
+gunicorn = "^22.0.0"
+orjson = "^3.10.6"
+pydantic-settings = "^2.3.4"
+pydantic = "^2.8.2"
 
 
 [tool.poetry.group.dev.dependencies]
-isort = "^5.12.0"
-black = "^23.9.0"
-mypy = "^1.5.1"
-pytest-asyncio = "^0.21.1"
-coverage = {extras = ["toml"], version = "^7.3.1"}
-pytest = "^7.4.2"
-pytest-cov = "^4.1.0"
-pytest-socket = "^0.6.0"
-pytest-mock = "^3.11.1"
-taskipy = "^1.12.0"
-ipython = "^8.15.0"
+mypy = "^1.10.1"
+pytest-asyncio = "^0.23.7"
+coverage = {extras = ["toml"], version = "^7.5.4"}
+pytest = "^8.2.2"
+pytest-cov = "^5.0.0"
+pytest-socket = "^0.7.0"
+pytest-mock = "^3.14.0"
+taskipy = "^1.13.0"
+ipython = "^8.26.0"
+ruff = "^0.5.1"
 
 [tool.taskipy.tasks]
-run-app-local = "uvicorn app.main:create_app --reload"
-format-and-lint = "task isort && task mypy-lint && task black"
+format-and-lint = "task ruff && task mypy-lint"
 
-isort = "isort app tests"
-tests = "pytest tests/ -v"
-black = "black app/ tests/"
+ruff = "ruff format ./app/ ./docker/ ./tests/ && ruff check ./app/ ./docker/ ./tests/ --fix --unsafe-fixes"
+tests = "PYTHONPATH=. pytest tests/ -v"
 
-isort-lint = "task isort -c"
-black-lint = "task black --check"
-mypy-lint = "mypy --install-types --non-interactive app/"
+ruff-lint = "ruff check ./app"
+mypy-lint = "mypy --install-types --non-interactive ./app/"
 poetry-check = "poetry check"
 pip-check = "pip check"
 
-[tool.isort]
-# isort configuration:
-# https://github.com/timothycrosley/isort/wiki/isort-Settings
-include_trailing_comma = true
-use_parentheses = true
-# See https://github.com/timothycrosley/isort#multi-line-output-modes
-multi_line_output = 3
-line_length = 120
-known_first_party = "app"
-default_section = "THIRDPARTY"
-sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
-
-
-[tool.black]
+[tool.ruff]
 line-length = 120
-target-version = ['py311']
-extend-exclude = '''
-(
-  /(
-    # Exclude folders
-    | app/migrations
-  )/
-  # Exclude files
-)
-'''
-skip-string-normalization = true
+extend-exclude = ["site-packages", "*.pyi"]
+target-version = 'py312'
+
+
+[tool.ruff.lint]
+fixable = ["ALL"]
+unfixable = []
+select = [
+    "E",  # pycodestyle errors
+    "W",  # pycodestyle warnings
+    "F",  # pyflakes
+    "I",  # isort
+    "C",  # flake8-comprehensions
+    "B",  # flake8-bugbear
+    "UP",  # pyupgrade
+]
+ignore = [
+    "B008",  # do not perform function calls in argument defaults
+    "C901",  # too complex
+]
+
+[tool.ruff.format]
+quote-style = 'single'
+indent-style = 'space'
+line-ending = 'auto'
+skip-magic-trailing-comma = false
+exclude = ['*.pyi']
+
+[tool.ruff.lint.mccabe]
+max-complexity = 6
+
+[tool.ruff.lint.isort]
+split-on-trailing-comma = false
+lines-after-imports = 2
+known-first-party = ["app"]
+
+[tool.ruff.lint.pyupgrade]
+# Preserve types, even if a file imports `from __future__ import annotations`.
+keep-runtime-typing = true
 
 
 [tool.mypy]
