@@ -4,5 +4,4 @@ from typing import Any
 
 class AbstractUseCase(ABC):
     @abstractmethod
-    async def execute(self) -> Any:
-        ...
+    async def execute(self) -> Any: ...
